+++ weblate/trans/templatetags/translations.py
@@ -541,16 +541,7 @@ def get_location_links(profile, unit):
         return gettext("string ID %s") % unit.location
 
     # Go through all locations separated by comma
-    for location in unit.location.split(","):
-        location = location.strip()
-        if location == "":
-            continue
-        location_parts = location.split(":")
-        if len(location_parts) == 2:
-            filename, line = location_parts
-        else:
-            filename = location_parts[0]
-            line = 0
+    for location, filename, line in unit.get_locations():
         link = unit.translation.component.get_repoweb_link(
             filename, line, profile.editor_link
         )