+++ docs/changes.rst
@@ -5,6 +5,7 @@ Not yet released.
 
 * Weblate now requires Python 3.5 or newer.
 * Added management overview of component alerts.
+* Added component alert for broken repository browser URLs.
 
 Weblate 3.11.3
 --------------